HH-218702 move DataSourceType to nab-jdbc-common
