+++ nab-data-source/src/main/java/ru/hh/nab/datasource/DataSourceContextUnsafe.java
@@ -4,6 +4,7 @@ import java.util.Optional;
 import static java.util.Optional.ofNullable;
 import java.util.function.Supplier;
 import ru.hh.nab.common.mdc.MDC;
+import ru.hh.nab.jdbc.common.DataSourceType;
 
 public final class DataSourceContextUnsafe {
   public static final String MDC_KEY = "db";
