+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/datasource/RoutingDataSource.java
@@ -14,8 +14,8 @@ import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.lang.Nullable;
 import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
-import ru.hh.nab.datasource.DataSourceType;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
+import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.common.datasource.NamedDataSource;
 import ru.hh.nab.jdbc.common.ext.JdbcExtension;
 import ru.hh.nab.jdbc.common.healthcheck.HealthCheck;
