+++ nab-tests/src/test/java/ru/hh/nab/datasource/DataSourceContextTransactionManagerTest.java
@@ -16,6 +16,7 @@ import static org.springframework.transaction.support.TransactionSynchronization
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isCurrentTransactionReadOnly;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
+import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jpa.JpaTestConfig;
 import ru.hh.nab.jpa.model.TestEntity;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
