+++ nab-tests/src/test/java/ru/hh/nab/datasource/DataSourceTestConfig.java
@@ -7,6 +7,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import ru.hh.nab.common.properties.FileSettings;
+import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.testbase.NabTestConfig;
 import ru.hh.nab.testbase.postgres.embedded.EmbeddedPostgresDataSourceFactory;
