+++ nab-tests/src/test/java/ru/hh/nab/jpa/DataSourceSwitchingTest.java
@@ -39,7 +39,6 @@ import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
 import static ru.hh.nab.datasource.DataSourceSettings.HEALTHCHECK_ENABLED;
 import static ru.hh.nab.datasource.DataSourceSettings.HEALTHCHECK_SETTINGS_PREFIX;
 import static ru.hh.nab.datasource.DataSourceSettings.ROUTING_SECONDARY_DATASOURCE;
-import ru.hh.nab.datasource.DataSourceType;
 import ru.hh.nab.datasource.DatabaseSwitcher;
 import ru.hh.nab.datasource.healthcheck.AsyncHealthCheck;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSource;
@@ -51,6 +50,7 @@ import ru.hh.nab.hibernate.datasource.RoutingDataSource;
 import ru.hh.nab.hibernate.datasource.RoutingDataSourceFactory;
 import static ru.hh.nab.hibernate.transaction.DataSourceContext.onDataSource;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
+import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.common.healthcheck.HealthCheckResult;
 import ru.hh.nab.jpa.model.TestEntity;
 import ru.hh.nab.metrics.StatsDSender;
