+++ nab-tests/src/test/java/ru/hh/nab/jpa/JpaTestConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import ru.hh.nab.common.properties.FileSettings;
 import ru.hh.nab.datasource.DataSourceFactory;
-import ru.hh.nab.datasource.DataSourceType;
+import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jpa.model.TestEntity;
 import ru.hh.nab.testbase.NabTestConfig;
 import ru.hh.nab.testbase.hibernate.NabHibernateTestBaseConfig;
